### 2024-07-26 #7041 @editorialbot
[PRE REVIEW]: EPyT-Flow: A Toolkit for Generating Water Distribution Network Data
**Submitting author:** @andreArtelt (<a href="http://orcid.org/0000-0002-2426-3126">André Artelt</a>)
**Repository:** https://github.com/WaterFutures/EPyT-Flow
**Branch with paper.md** (empty if default branch): dev
**Version:** v0.6.0
