### 2024-07-27 d0337a3
Add parameter to place_pressure_sensors_everywhere function
