### 2024-08-05 e3d631d
Upgrade EPyT version -- warnings can now be caught
