+++ REQUIREMENTS.txt
@@ -1,4 +1,4 @@
-epyt>=1.1.9
+epyt>=1.2.0
 requests>=2.31.0
 scipy>=1.11.4
 u-msgpack-python>=2.8.0
