### 2024-08-08 50066ac
Add sanity checks: some functions can not be called when simulation is running
