+++ epyt_flow/simulation/scenario_simulator.py
@@ -94,6 +94,7 @@ class ScenarioSimulator():
         self.__controls = []
         self.__system_events = []
         self.__sensor_reading_events = []
+        self.__running_simulation = False
 
         custom_epanet_lib = None
         custom_epanetmsx_lib = None
@@ -875,6 +876,9 @@ class ScenarioSimulator():
         """
         Randomizes all demand patterns.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("Can not change general parameters when simulation is running.")
+
         self.__adapt_to_network_changes()
 
         # Get all demand patterns
@@ -997,6 +1001,9 @@ class ScenarioSimulator():
         event : :class:`~epyt_flow.simulation.events.system_event.SystemEvent`
             System event.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("Can not add events when simulation is running.")
+
         self.__adapt_to_network_changes()
 
         if not isinstance(event, SystemEvent):
@@ -1016,6 +1023,9 @@ class ScenarioSimulator():
         sensor_fault_event : :class:`~epyt_flow.simulation.events.sensor_faults.SensorFault`
             Sensor fault specifications.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("Can not add events when simulation is running.")
+
         self.__adapt_to_network_changes()
 
         sensor_fault_event.validate(self.__sensor_config)
@@ -1036,6 +1046,9 @@ class ScenarioSimulator():
         sensor_reading_attack : :class:`~epyt_flow.simulation.events.sensor_reading_attack.SensorReadingAttack`
             Sensor fault specifications.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("Can not add events when simulation is running.")
+
         self.__adapt_to_network_changes()
 
         sensor_reading_attack.validate(self.__sensor_config)
@@ -1056,6 +1069,9 @@ class ScenarioSimulator():
         event : :class:`~epyt_flow.simulation.events.sensor_reading_event.SensorReadingEvent`
             Sensor reading event.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("Can not add events when simulation is running.")
+
         self.__adapt_to_network_changes()
 
         event.validate(self.__sensor_config)
@@ -1453,6 +1469,9 @@ class ScenarioSimulator():
         :class:`~epyt_flow.simulation.scada.scada_data.ScadaData`
             Quality simulation results as SCADA data.
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("A simulation is already running.")
+
         if self.__f_msx_in is None:
             raise ValueError("No .msx file specified")
 
@@ -1517,6 +1536,9 @@ class ScenarioSimulator():
             Generator containing the current EPANET-MSX simulation results as SCADA data
             (i.e. species concentrations).
         """
+        if self.__running_simulation is True:
+            raise RuntimeError("A simulation is already running.")
+
         if self.__f_msx_in is None:
             raise ValueError("No .msx file specified")
 
@@ -1532,6 +1554,8 @@ class ScenarioSimulator():
         hyd_time_step = self.epanet_api.getTimeHydraulicStep()
 
         self.epanet_api.initializeMSXQualityAnalysis(ToolkitConstants.EN_NOSAVE)
+
+        self.__running_simulation = True
 
         bulk_species_idx = self.epanet_api.getMSXSpeciesIndex(self.__sensor_config.bulk_species)
         surface_species_idx = self.epanet_api.getMSXSpeciesIndex(
@@ -1667,6 +1691,8 @@ class ScenarioSimulator():
                                         sensor_noise=self.__sensor_noise,
                                         frozen_sensor_config=frozen_sensor_config)
 
[…]
