### 2024-08-15 #7104 @andreArtelt
Thanks a lot, Krik and Meghna for agreeing to review my submission. I think you both have a great amount of knowledge and experience in the water domain. Looking forward to a constructive review process :)

A quick note from my side: From tomorrow on, I will be on vacation for 6 weeks with only very limited internet access. However, my co-author  @Mariosmsk will keep an eye on this issue as well as on the EPyT-Flow repository.

Best,
André
