### 2024-09-06 WaterFutures/EPyT-Flow#9 @meghnathomas
JOSS Review
Hi! I will be conducting my JOSS review in this issue.

epyt-flow is a really interesting and useful package that offers a large range of functions and very comprehensive documentation! The inclusion of several benchmark networks and scenarios from the literature is a great touch and will aid in adoption by water distribution systems (WDS) researchers. I was able to successfully install the package (using pip) and run all examples. Below I have included comments and questions on different aspects of the package, repository, and docs.

General

- [x] It was a little unclear to me how SCADA data is generated. Is SCADA data basically = pressure/flow simulations results for the nodes/links at which sensors are installed? So if we run a simulation with no uncertainties and no events, will SCADA data results be equal to the results of a basic hydraulic simulation as performed by EpyT or WNTR?
- [ ] If I have a network model of a real WDS + SCADA data collected from meters placed throughout the system, I see from Docs -> In-depth Tutorial -> SCADA data that I can use epyt-flow to analyze the model and integrate the data. Would you be able to add an example demonstrating this (perhaps using a common network such as C-TOWN or BWFLNet)?
- [x] Can epyt-flow read controls from the .INP file of a model or must a user define an AdvancedControlModule class to specify controls (like those shown in the Custom Controls example)?
- [x] #16

Landing page

- [x] I recommend adding a line of code to plot a timeseries of the pressure and flow SCADA data results + the actual plots themselves for the Quick Example
- [x] #14

Docs: in-depth tutorial

- [x] #22
- [x] #12
- [x] #11
- [x] #20
- [x] Control: I see that in the example you set the tank diameter by calling `self.__tank_diameter = 50.5`. Is it possible to change all system parameters (e.g., pipe diameter, node elevation) in this way?
- [x] Control: EPANET offers simple time based controls, simple event based controls, and rule based controls. Can we use epyt-flow to model rule-based controls as well?
- [x] #13
- [x] #15
- [x] #21

Docs: examples

- [x] #18
- [x] Working with Network Topology: when we list the edges with `print(topo.edges)`, why is the second element in each tuple a string?
- [x] Uncertainties Example: I am a huge fan of epyt-flow's ability to add uncertainties to model parameters! What parameters can one add uncertainty to besides demand patterns? And can we add different kinds of uncertainty to, say, different demand patterns or base demands or must all instances of a parameter be assigned the same uncertainty?
- [x] #19
- [x] Water Age Analysis and Chloring Injection: would the results for these two examples be the same as in EPANET?
- [x] #17

Examples
- [x] I recommend adding the plotting functions from the docs into the Python files so users can immediately generate and reproduce your plots and do not have to refer to the docs (like for the event detection and net2-cl2 examples)

Paper

- [x] You mention that WNTR does not consider industrial controls, but I believe it offers the same control functionalities that epyt-flow does (simple event- and time-based controls) as well as control rules!
