### 2024-09-07 WaterFutures/EPyT-Flow#9 @andreArtelt
Thanks a lot for this detailed and constructive feedback - much appreciated 🙏

I am currently on vacation until 5th October - I will work on those aspects once I am back in office.
Thanks for your understanding.

Best,
André
