### 2024-10-09 WaterFutures/EPyT-Flow#9 @andreArtelt
I was busy with administrative tasks the last days after I returned from a longer vacation -- sorry about the delay.

Regarding your comment on the paper: I think I did not make this clear enough in the paper. EPyT-Flow supports arbitrary controls that go beyond "simple" IF-THEN-ELSE rules as supported by EPANET. For instance, you could use a neural network to make control decisions based on the observed sensor readings. I added a clarifying statement to the paper and documentation in commit dee0de5
Let me know, what you think.

Also, how do you want me to work on your comments/tasks. There are quite a few of them and I was wondering whether it would make sense to convert some of them into issues to have a better overview.
