### 2024-10-09 WaterFutures/EPyT-Flow#10 @andreArtelt
Your comment on controls is very similar to the one raised by the other reviewer in #9
I think I did not make this clear enough in the paper. EPyT-Flow supports arbitrary controls that go beyond "simple" IF-THEN-ELSE rules as supported by EPANET. For instance, you could use a neural network to make control decisions based on the observed sensor readings. I added a clarifying statement to the paper and documentation in commit dee0de5
Let me know, what you think.
