### 2024-10-09 WaterFutures/EPyT-Flow#10 @andreArtelt
Regarding your comment on missing quality dynamics in WNTR:
You are right that WNTR provides access to EPANET quality dynamics and hopefully to EPANET-MSX as well soon. However, if I understood it correctly, those simulators can not be used with the other (custom) simulator and therefore lack some features such as the effect of leakages, earth quakes, etc.. This is not the case in EPyT-Flow, here everything is integrated into a single simulator -- e.g. leakages and quality dynamics can be modeled together.

My suggestion would be to relax and rephrase the statement in the paper, e.g. "does not fully integrate (advanced) quality dynamics" or smth. similar. What do you think?
