### 2024-10-09 efa61b5
#10 Add statement of need to landing page of documentation
