### 2024-10-09 be17bda
#10 Add issue templates
