+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,22 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**What is the current behavior?**
+
+**If the current behavior is a bug, please provide the steps to reproduce and if possible a minimal running code example of the problem**
+
+**What is the expected behavior?**
+
+**What is the motivation / use case for changing the behavior?**
+
+**Please tell us about your environment:**
+
+Version:
+Python version:
+OS:
