### 2024-10-09 cb820c0
#10 README: Add short statement on how to get support
