### 2024-10-10 9f2c895
#9 Add plots to quick example
