+++ README.md
@@ -86,7 +86,7 @@ pip install .
 ```python
 from epyt_flow.data.benchmarks import load_leakdb_scenarios
 from epyt_flow.simulation import ScenarioSimulator
-from epyt_flow.utils import to_seconds
+from epyt_flow.utils import to_seconds, plot_timeseries_data
 
 
 if __name__ == "__main__":
@@ -107,10 +107,20 @@ if __name__ == "__main__":
         # Run entire simulation
         scada_data = sim.run_simulation()
 
-        # Show sensor readings over the entire simulation
+        # Print & plot sensor readings over the entire simulation
         print(f"Pressure readings: {scada_data.get_data_pressures()}")
+        plot_timeseries_data(scada_data.get_data_pressures().T,
+                             x_axis_label="Time (30min steps)",
+                             y_axis_label="Pressure in $m$")
+
         print(f"Flow readings: {scada_data.get_data_flows()}")
+        plot_timeseries_data(scada_data.get_data_flows().T,
+                             x_axis_label="Time (30min steps)",
+                             y_axis_label="Flow rate in $m^3/h$")
 ```
+<img src="https://github.com/WaterFutures/EPyT-Flow/blob/dev/docs/_static/examples_basic_usage_pressure.png?raw=true" align="left" width="49%"/>
+<img src="https://github.com/WaterFutures/EPyT-Flow/blob/dev/docs/_static/examples_basic_usage_flow.png?raw=true" align="right" width="49%"/>
+
 
 ## Documentation
 
