### 2024-10-10 WaterFutures/EPyT-Flow#10 @andreArtelt
> > Right now, the statement of need is only in the paper. However, in [efa61b5](https://github.com/WaterFutures/EPyT-Flow/commit/efa61b57819158e835eeec74ceb79f4da90efecc) I just added it to the landing page of the documentation as well.
>
> Thanks. That placement looks good, but I haven't seen it update on the actual documentation page. Not sure what triggers the docs to be updated, perhaps it is only on releases?

Yes, the documentation is only rebuilt in case of a new release. However, you can check-out the documentation of the dev branch on [https://epyt-flow.readthedocs.io/en/latest/](https://epyt-flow.readthedocs.io/en/latest/) -- I just triggered the built of this one manually, so that you can see how it will look like if I make a new release.
