### 2024-10-10 a675a15
#9 Examples: Add plot functions from Jupyter notebooks
