+++ examples/sensor_fault.py
@@ -4,7 +4,7 @@ Example of implementing a sensor fault.
 from epyt_flow.data.networks import load_hanoi
 from epyt_flow.simulation import ScenarioSimulator, SENSOR_TYPE_NODE_PRESSURE
 from epyt_flow.simulation.events import SensorFaultStuckZero
-from epyt_flow.utils import to_seconds
+from epyt_flow.utils import to_seconds, plot_timeseries_data
 
 
 if __name__ == "__main__":
@@ -29,3 +29,5 @@ if __name__ == "__main__":
 
         # Retrieve pressure readings at node "16"
         print(scada_data.get_data_pressures(sensor_locations=["16"]))
+        plot_timeseries_data(scada_data.get_data_pressures(sensor_locations=["16"]).T,
+                             x_axis_label="Time (30min steps)", y_axis_label="Pressure in $m$")
