+++ examples/quick_example.py
@@ -4,7 +4,7 @@ Minimalistic usage example of EPyt-Flow.
 from epyt_flow.simulation.scada.scada_data_export import ScadaDataXlsxExport
 from epyt_flow.data.benchmarks import load_leakdb_scenarios
 from epyt_flow.simulation import ScenarioSimulator
-from epyt_flow.utils import to_seconds
+from epyt_flow.utils import to_seconds, plot_timeseries_data
 
 
 if __name__ == "__main__":
@@ -28,6 +28,11 @@ if __name__ == "__main__":
         # Export data in excel
         ScadaDataXlsxExport(f_out="myHanoiResults.xlsx").export(scada_data)
 
-        # Show sensor readings over the entire simulation
+        # Print & plot sensor readings over the entire simulation
         print(f"Pressure readings: {scada_data.get_data_pressures()}")
         print(f"Flow readings: {scada_data.get_data_flows()}")
+
+        plot_timeseries_data(scada_data.get_data_pressures().T,
+                             x_axis_label="Time (30min steps)", y_axis_label="Pressure in $m$")
+        plot_timeseries_data(scada_data.get_data_flows().T,
+                             x_axis_label="Time (30min steps)", y_axis_label="Flow rate in $m^3/h$")
