+++ examples/pump_states.py
@@ -3,7 +3,7 @@ Example of setting the state of a pump in an acutator event.
 """
 from epyt_flow.data.networks import load_net1
 from epyt_flow.simulation import ScenarioSimulator
-from epyt_flow.utils import to_seconds
+from epyt_flow.utils import to_seconds, plot_timeseries_data
 from epyt_flow.simulation.events import PumpStateEvent, ActuatorConstants
 
 
@@ -34,3 +34,9 @@ if __name__ == "__main__":
         # Run simulation and show sensor readings over time
         scada_data = sim.run_simulation()
         print(scada_data.get_data())
+
+        plot_timeseries_data(scada_data.get_data_pumps_state().T,
+                             x_axis_label="Time (30min steps)", y_axis_label="Pump state")
+        plot_timeseries_data(scada_data.get_data_tanks_water_volume().T,
+                             x_axis_label="Time (30min steps)",
+                             y_axis_label="Water volume in $m^3$")
