+++ examples/uncertainties.py
@@ -4,7 +4,7 @@ Example of adding uncertainty with respect to WDN parameters (e.g. demand patter
 import numpy as np
 from epyt_flow.data.networks import load_ltown
 from epyt_flow.simulation import ScenarioSimulator, ModelUncertainty, RelativeUniformUncertainty
-from epyt_flow.utils import to_seconds
+from epyt_flow.utils import to_seconds, plot_timeseries_data
 
 
 if __name__ == "__main__":
@@ -30,3 +30,5 @@ if __name__ == "__main__":
             measurements.append(scada_data.get_data_pressures(sensor_locations=["n105"]).
                                 flatten().tolist())
         print(np.mean(measurements, axis=0), np.var(measurements, axis=0))
+        plot_timeseries_data(np.array(measurements),
+                             x_axis_label="Time (5min steps)", y_axis_label="Pressure in $m$")
