### 2024-10-10 d3ce527
Fix image formatting in README.md
