### 2024-10-10 WaterFutures/EPyT-Flow#9 @meghnathomas
Hi @andreArtelt  sorry for the late reply! That works for me - would you like me to split this comment up into separate issues or would you prefer to do that yourself?
