### 2024-10-11 WaterFutures/EPyT-Flow#9 @andreArtelt
> Hi @andreArtelt sorry for the late reply! That works for me - would you like me to split this comment up into separate issues or would you prefer to do that yourself?

I can do that myself. I think I might just reply to some of your comments in this issue if your comment is more a question than an issue that requires changes (i.e. commits).
