### 2024-10-11 WaterFutures/EPyT-Flow#10 @andreArtelt
> When I run the test suite all tests pass except for the benchmark tests which seem to get caught up on some request issues. I'll work on this a little bit more but if I can't resolve it I'll open a separate issue about the problem.

Does this issue still exist? All tests run successfully on my machine as well as on GitHub (see [Actions](https://github.com/WaterFutures/EPyT-Flow/actions)) -- all tests are run when a commit is pushed.
