### 2024-10-11 WaterFutures/EPyT-Flow#9 @andreArtelt
> I recommend adding a line of code to plot a timeseries of the pressure and flow SCADA data results + the actual plots themselves for the Quick Example

Was already addressed in [9f2c895](https://github.com/WaterFutures/EPyT-Flow/commit/9f2c8951487a433d439b1da9b2803cec1f75e92b)
