### 2024-10-11 WaterFutures/EPyT-Flow#9 @andreArtelt
> You mention that WNTR does not consider industrial controls, but I believe it offers the same control functionalities that epyt-flow does (simple event- and time-based controls) as well as control rules!

Was addressed in one of my previous comments and in [dee0de5](https://github.com/WaterFutures/EPyT-Flow/commit/dee0de592b8f3eb1fcf82eadc054d198e44ca398)
