### 2024-10-11 9eccac1
#11 Docs: Explain why it is important to close the scenario
