### 2024-10-11 98c95f7
#14 README: Add "How to Get Started?" section
