### 2024-10-11 d9bf720
#13 Doc: In-depth explanation of leakage modeling and implementation
