### 2024-10-11 a5a541f
#15 Doc: Information about order of applying controls and events
