### 2024-10-11 WaterFutures/EPyT-Flow#9 @andreArtelt
> Water Age Analysis and Chloring Injection: would the results for these two examples be the same as in EPANET?

Yes, the results would be exactly the same because EPyT-Flow uses EPANET for the simulation.
