### 2024-10-11 887e27b
Quick example: Add plot functions from Jupyter notebooks
