### 2024-10-12 4a83800
Plot functions: User can specify y ticks
