### 2024-10-12 c74955e
#16 Doc: Add a brief statement on the default sensor configurations
