### 2024-10-12 8f2f55b
Bugfix: Add missing SI unit to is_flowunit_simetric
