### 2024-10-13 392f843
#18 Docs: Add legends to line plots and use more descriptive y ticks for pumps
