+++ README.md
@@ -110,6 +110,8 @@ if __name__ == "__main__":
         # Print & plot sensor readings over the entire simulation
         print(f"Pressure readings: {scada_data.get_data_pressures()}")
         plot_timeseries_data(scada_data.get_data_pressures().T,
+                             labels=[f"Node {n_id}" for n_id in
+                                     scada_data.sensor_config.pressure_sensors],
                              x_axis_label="Time (30min steps)",
                              y_axis_label="Pressure in $m$")
 
