### 2024-10-13 WaterFutures/EPyT-Flow#9 @andreArtelt
> Can epyt-flow read controls from the .INP file of a model or must a user define an AdvancedControlModule class to specify controls (like those shown in the Custom Controls example)?

Right now,  controls from .INP files are not parsed. I think, this smth. for a future release.
