### 2024-10-14 9da8421
#19 Abrupt Leakage Example: Revise plot
