### 2024-10-14 WaterFutures/EPyT-Flow#9 @meghnathomas
> > Working with Network Topology: when we list the edges with print(topo.edges), why is the second element in each tuple a string?
>
> Every element in each tuple should be a string -- if you observe any other behavior, please create a bug report.
>
> The result of `topo.edges` is a list of tuples (edges) -- each edge is represented by the IDs (strings) of the two nodes which it is connecting.

Ok I see - I think I was confused because in your [docs ](https://epyt-flow.readthedocs.io/en/stable/examples/network_topology.html)(code block 7), it looks like the first element of the tuple is a regular string while the second element is a numpy.string_. Not a big deal!
