### 2024-10-15 9d1df3f
#20 Doc: Add details about the source of potential exceptions
