### 2024-10-15 WaterFutures/EPyT-Flow#9 @andreArtelt
> Control: I see that in the example you set the tank diameter by calling self.__tank_diameter = 50.5. Is it possible to change all system parameters (e.g., pipe diameter, node elevation) in this way?

This does not change any EPANET parameters. This is just a local variable that is used in the control rule. As mentioned in the comment (in the notebook), the value could have been obtained by calling an EPANET function as well. However, changing this local variable has no effect on the actual tank diameter. The tank diameter can be changed by either editing the .inp file or by calling EPyT functions such as `setNodeTankDiameter`
