### 2024-10-15 70a3b5d
#21 Doc: Emphasize that multiple sensor faults can act on the same sensor
