### 2024-10-15 WaterFutures/EPyT-Flow#9 @andreArtelt
> Uncertainties Example: I am a huge fan of epyt-flow's ability to add uncertainties to model parameters! What parameters can one add uncertainty to besides demand patterns? And can we add different kinds of uncertainty to, say, different demand patterns or base demands or must all instances of a parameter be assigned the same uncertainty?

Uncertainties can be added to all parameters which are listed in the documentation of the `ModelUncertainty` class (see https://epyt-flow.readthedocs.io/en/stable/epyt_flow.uncertainty.html#epyt_flow.uncertainty.model_uncertainty.ModelUncertainty)

Right now, only "global" uncertainties can be added -- e.g. one type of uncertainty for all base demands. Being able to be more selective (e.g. having different kinds of uncertainties for different base demands) is a great idea and will be added to my todo list for future releases of EPyT-Flow :)
