### 2024-10-15 WaterFutures/EPyT-Flow#9 @andreArtelt
> It was a little unclear to me how SCADA data is generated. Is SCADA data basically = pressure/flow simulations results for the nodes/links at which sensors are installed? So if we run a simulation with no uncertainties and no events, will SCADA data results be equal to the results of a basic hydraulic simulation as performed by EpyT or WNTR?

Yes, SCADA data refers to the simulation results (pressure, flows, species concentrations, pump states, etc.) that are accessible to the user by placing sensors. However, those SCADA data (i.e. sensor readings) might be affected by sensor faults, uncertainties, etc. Without any uncertainties and sensor faults, the results are the same as in EPyT/EPANET and should not differ much from WNTR (be aware of the units!).
