### 2024-10-15 WaterFutures/EPyT-Flow#9 @andreArtelt
> If I have a network model of a real WDS + SCADA data collected from meters placed throughout the system, I see from Docs -> In-depth Tutorial -> SCADA data that I can use epyt-flow to analyze the model and integrate the data. Would you be able to add an example demonstrating this (perhaps using a common network such as C-TOWN or BWFLNet)?

I am not sure if I understand you correctly. Do you want to load real SCADA data into EPyT-Flow? If so, this could be done by manually creating an instance of `ScadaData` (see https://epyt-flow.readthedocs.io/en/stable/epyt_flow.simulation.scada.html#epyt_flow.simulation.scada.scada_data.ScadaData) -- however, I am not sure what you want to analyze or integrate in this context.
