+++ docs/tut.scada.rst
@@ -321,6 +321,8 @@ Example of getting and changing the measurement units:
     print(flowunit_to_str(scada_data_new.sensor_config.flow_unit))
 
 
+.. _scada_import_export:
+
 Importing and Exporting
 +++++++++++++++++++++++
 
