### 2024-10-22 0d0b1da
running_mse: Add type hints of return value
