### 2024-10-23 07fa920
Add wrappers for sklearn regression metrics
