### 2024-10-24 baf8677
Plot functions: Add option to save plot to file
