### 2024-10-25 7230780
Bugfix: Serialization of PercentageDeviationUncertainty
