### 2024-10-26 c8aa6f8
Serialization: Add functions for loading and saving JSON files
