### 2024-10-27 WaterFutures/EPyT-Flow#10 @kbonney
> > When I run the test suite all tests pass except for the benchmark tests which seem to get caught up on some request issues. I'll work on this a little bit more but if I can't resolve it I'll open a separate issue about the problem.
>
> Does this issue still exist? All tests run successfully on my machine as well as on GitHub (see [Actions](https://github.com/WaterFutures/EPyT-Flow/actions)) -- all tests are run when a commit is pushed.

Hi @andreArtelt, I reinstalled the package on my personal machine and was able to get everything passing. Sometimes the VPN I use for work gets in the way of requests in Python, so the problem probably isn't on the package's end. I'll mark this complete.
