### 2024-10-27 WaterFutures/EPyT-Flow#10 @kbonney
> > > Your comment on controls is very similar to the one raised by the other reviewer in #9 I think I did not make this clear enough in the paper. EPyT-Flow supports arbitrary controls that go beyond "simple" IF-THEN-ELSE rules as supported by EPANET. For instance, you could use a neural network to make control decisions based on the observed sensor readings. I added a clarifying statement to the paper and documentation in commit [dee0de5](https://github.com/WaterFutures/EPyT-Flow/commit/dee0de592b8f3eb1fcf82eadc054d198e44ca398) Let me know, what you think.
> >
> >
> > Thanks, that helps clarify the differences. It is also nice to highlight the use of ML throughout the paper, as I think that aspect is a bit understated.
> > > Regarding your comment on missing quality dynamics in WNTR: You are right that WNTR provides access to EPANET quality dynamics and hopefully to EPANET-MSX as well soon. However, if I understood it correctly, those simulators can not be used with the other (custom) simulator and therefore lack some features such as the effect of leakages, earth quakes, etc.. This is not the case in EPyT-Flow, here everything is integrated into a single simulator -- e.g. leakages and quality dynamics can be modeled together.
> > > My suggestion would be to relax and rephrase the statement in the paper, e.g. "does not fully integrate (advanced) quality dynamics" or smth. similar. What do you think?
> >
> >
> > Yes, the leaks in the custom simulator can't be used with the EPANET simulator. I think in your rephrasing I would mention the specific cases you brought up. "does not fully integrate (advanced) quality dynamics with scenarios such as pipe leaks."
>
> Done in [21b9344](https://github.com/WaterFutures/EPyT-Flow/commit/21b934473f177b5885b43d2a69281208d6637326) -- I also added a few sentenced highlighting the benefits for AI and ML research

Looks good.

> > > Right now, the statement of need is only in the paper. However, in [efa61b5](https://github.com/WaterFutures/EPyT-Flow/commit/efa61b57819158e835eeec74ceb79f4da90efecc) I just added it to the landing page of the documentation as well.
> >
> >
> > Thanks. That placement looks good, but I haven't seen it update on the actual documentation page. Not sure what triggers the docs to be updated, perhaps it is only on releases?
>
> Yes, the documentation is only rebuilt in case of a new release. However, you can check out the documentation of the dev branch on https://epyt-flow.readthedocs.io/en/latest/ -- I just triggered the build of this one manually, so that you can see how it will look like if I make a new release.

Looks good, thanks.
