### 2024-10-28 ed5d3e3
Bugfix: NetworkTopology correctly compare links
