### 2024-10-29 0f03448
Add new JsonSerializable functions to documentation
