### 2024-10-30 693b3d5
ScenarioSimulator: Add function get_reporting_time_step
