### 2024-10-30 22de93a
Bugfix: Correctly remove temporary .msx file
