### 2024-10-31 0c39a52
Bugfix in bulk_species_node_concentration
