### 2024-11-05 a793ae1
ScenarioSimulator: Add function add_species_injection_source for EPANET-MSX species
