### 2024-11-05 4bd0481
Bugfix: Correct time pattern step in LeakDB
