### 2024-11-05 2a94b69
Added extended visualization
Added extended visualization functionality
