### 2024-11-05 e06e066
Update REQUIREMENTS.txt
