+++ REQUIREMENTS.txt
@@ -10,3 +10,5 @@ falcon>=3.1.3
 multiprocess>=0.70.16
 geopandas>=0.14.4
 psutil
+matplotlib
+svgpath2mpl
