### 2024-11-06 WaterFutures/EPyT-Flow#9 @meghnathomas
> > If I have a network model of a real WDS + SCADA data collected from meters placed throughout the system, I see from Docs -> In-depth Tutorial -> SCADA data that I can use epyt-flow to analyze the model and integrate the data. Would you be able to add an example demonstrating this (perhaps using a common network such as C-TOWN or BWFLNet)?
>
> I am not sure if I understand you correctly. Do you want to load real SCADA data into EPyT-Flow? If so, this could be done by manually creating an instance of `ScadaData` (see https://epyt-flow.readthedocs.io/en/stable/epyt_flow.simulation.scada.html#epyt_flow.simulation.scada.scada_data.ScadaData) -- however, I am not sure what you want to analyze or integrate in this context.

This feature would be very useful in calibration and state estimation tasks where we want to use information from both the hydraulic simulation results and sparse SCADA data to correct parameters (like pipe roughnesses) or estimate real-time system-wide pressures and flows. There are different methods to actually perform the state estimation/calibration procedures, and I am not concerned about the implementation of those - just wondering how one could use EpyT-Flow to access a hydraulic model of their own network and use the function in the link you provided to bring in SCADA data! Does that make sense?

I'm asking this because I have been trying to test a simple example to see how EpyT-Flow could be used to load a custom network that is not built into your collection of networks/scenarios. I am trying this out with CTown (I know CTown is in your database but I am trying to load the network as an INP file and load in [sensor data](http://www.water-simulation.com/wsp/about/bwcn/) using numpy arrays). At this point I am not concerned with the analysis, I am just pretending I am a user trying to load my own network and data so I can do my own analyses later. Can you walk me through that procedure?  E.g., to manually create an instance of ScadaData, it looks like I have to build a new sensor config. When I build a new [sensor config](https://epyt-flow.readthedocs.io/en/stable/epyt_flow.simulation.html#epyt_flow.simulation.sensor_config.SensorConfig), how can I access the list of nodes and links etc from the network INP file? It is a little challenging to navigate the EpyT-Flow code to work with networks/data that are not loaded into the package, so any guidance would be appreciated! And please let me know if this question or comment doesn't make sense I am happy to elaborate further
