### 2024-11-06 588e1b0
Add plot functions to ScadaData
