+++ epyt_flow/simulation/scada/scada_data.py
@@ -5,9 +5,11 @@ import warnings
 from typing import Callable, Any
 from copy import deepcopy
 import numpy as np
+import matplotlib
 from epyt.epanet import ToolkitConstants
 
-from ..sensor_config import SensorConfig, is_flowunit_simetric, massunit_to_str, \
+from ..sensor_config import SensorConfig, is_flowunit_simetric, massunit_to_str, flowunit_to_str,\
+    qualityunit_to_str, areaunit_to_str,\
     MASS_UNIT_MG, MASS_UNIT_UG, TIME_UNIT_HRS, MASS_UNIT_MOL, MASS_UNIT_MMOL, \
     AREA_UNIT_CM2, AREA_UNIT_FT2, AREA_UNIT_M2, \
     SENSOR_TYPE_LINK_FLOW, SENSOR_TYPE_LINK_QUALITY,  SENSOR_TYPE_NODE_DEMAND, \
@@ -18,6 +20,7 @@ from ..sensor_config import SensorConfig, is_flowunit_simetric, massunit_to_str,
 from ..events import SensorFault, SensorReadingAttack, SensorReadingEvent
 from ...uncertainty import SensorNoise
 from ...serialization import serializable, Serializable, SCADA_DATA_ID
+from ...utils import plot_timeseries_data
 
 
 @serializable(SCADA_DATA_ID, ".epytflow_scada_data")
@@ -1684,6 +1687,19 @@ class ScadaData(Serializable):
 
         return sensor_readings
 
+    def __get_x_axis_label(self) -> str:
+        if len(self.__sensor_readings_time) > 1:
+            time_step = self.__sensor_readings_time[1] - self.__sensor_readings_time[0]
+            if time_step > 60:
+                time_steps_desc = f"{int(time_step / 60)}min"
+                if time_step > 60*60:
+                    time_steps_desc = f"{int(time_step / 60)}hr"
+            else:
+                time_steps_desc = f"{time_step}s"
+            return f"Time ({time_steps_desc} steps)"
+        else:
+            return "Time"
+
     def get_data_pressures(self, sensor_locations: list[str] = None) -> np.ndarray:
         """
         Gets the final pressure sensor readings -- note that those might be subject to
@@ -1721,6 +1737,55 @@ class ScadaData(Serializable):
         idx = [self.__sensor_config.get_index_of_reading(pressure_sensor=s_id)
                for s_id in sensor_locations]
         return self.__sensor_readings[:, idx]
+
+    def plot_pressures(self, sensor_locations: list[str] = None, show: bool = True,
+                       save_to_file: str = None, ax: matplotlib.axes.Axes = None
+                       ) -> matplotlib.axes.Axes:
+        """
+        Plots the final pressure sensor readings -- note that those might be subject to
+        given sensor faults and sensor noise/uncertainty.
+
+        Parameters
+        ----------
+        sensor_locations : `list[str]`, optional
+            Existing pressure sensor locations for which the sensor readings have to be plotted.
+            If None, the readings from all pressure sensors are plotted.
+
+            The default is None.
+        show : `bool`, optional
+            If True, the plot/figure is shown in a window.
+
+            Only considered when 'ax' is None.
+
+            The default is True.
+        save_to_file : `str`, optional
+            File to which the plot is saved.
+
+            If specified, 'show' must be set to False --
+            i.e. a plot can not be shown and saved to a file at the same time!
+
+            The default is None.
+        ax : `matplotlib.axes.Axes`, optional
+            If not None, 'ax' is used for plotting.
+
+            The default is None.
+
+        Returns
+        -------
+        `matplotlib.axes.Axes`
+            Plot.
+        """
[…]
