### 2024-11-06 648bc4c
SensorConfig: Add function for getting species mass unit
