### 2024-11-06 WaterFutures/EPyT-Flow#9 @andreArtelt
Okay, I think I understood what you wanted to do. This is an interesting application of EPyT-Flow that I have not considered before. So far, I have always used (and built) EPyT-Flow to generate data but not to load external data (this is why there are export mechanisms for SCADA data but not import mechanisms).

Regarding your use case, here is how I would do it (I have not tested this -- there might be some typos xD):

```python
# Load .inp file
ScenarioSimulator(f_inp_in="CTown.inp") as scenario:
	  # 1. Build a sensor placement that is consistent with the data that you want to load and analyze
          # You could also do this after running the simulation if you whish to have different sensor configurations for the external data and the simulated data
	  scenario.set_pressure_sensors(sensor_locations=[...])
	  sim.set_flow_sensors(sensor_locations=[...])
	  # ...

	  # 2. Export this sensor config
	  my_sensor_config = sim.sensor_config()

          # 3. Run simulation
          scada_data = scenario.run_simulation()

	  # 4. Load external SCADA data -- everything must be a Numpy array!
	  my_measurement_time_points = ...
	  my_pressure_data = ...
	  my_flow_data = ...
	  # ...

	  # 5. Create a SCADA data instance containing the external
	  my_scada_data = ScadaData(sensor_config=my_sensor_config,
			            frozen_sensor_config=True,	# Assuming that the sensor config can not be changed -- set this to False if you loaded data for every node and link in the network!
                                    sensor_readings_time=my_measurement_time_points,
                                    pressure_data_raw=my_pressure_data
                                    flow_data_raw=my_flow_data
                                    ...)

	  # 6. Analyze the SCADA data using EPyT-Flow -- e.g. plotting data, adding sensor noise, comparing to the simulated data, etc.

	  # Plot pressures at all nodes over time
	  plot_timeseries_data(my_scada_data.get_data_pressures().T,
			       labels=[f"Node {n_id}" for n_id in my_scada_data.sensor_config.pressure_sensors],
			       x_axis_label="Time steps", y_axis_label="Pressure")

	  # ...
```

Let me know if this is what you have in mind.
