### 2024-11-07 a0e1915
ScadaDataExport: Bugfix MSX species
