### 2024-11-08 WaterFutures/EPyT-Flow#9 @meghnathomas
Thank you for the fast response, I appreciate how quickly you've addressed all my questions! :-) Yes - I think this is exactly what I was aiming for.  Would you be able to include a working example of this in your docs/examples for users to import their own networks and data? I tried to play with this rough code snippet a little bit and am confused about the format of the pressure data - you mention here that the input should be a numpy array, so I assumed it to be a 1D array containing my time-varying pressure data points, but the API docs say the pressure data input is "Raw pressure values of all nodes as a two-dimensional array – first dimension encodes time, second dimension pressure at nodes." Is that a 2 x T array (where T is the number of time steps) where the first row contains the time steps and second row contains the measurements? In any case,  using the 1D or 2D arrays using this code:

```
with ScenarioSimulator(f_inp_in='CTown.inp') as sim:
    sim.set_pressure_sensors(sensor_locations=['T1'])

    my_sensor_config = sim.sensor_config
    scada_data = sim.run_simulation()

    my_measurement_time_points = np.arange(0,3600*24, 3600)
    one_d_pressure_data = np.array([3, 2.82, 2.7, 2.62, 2.7, 2.89, 3.14, 3.26,
                                      3.4, 3.66, 3.73, 3.66, 3.73, 3.88, 4.07,
                                      4.23, 4.41, 4.44, 4.03, 4.03, 4.03, 4.03,
                                      4.03, 4.03])
    two_d_pressure_data = np.vstack([my_measurement_time_points, one_d_pressure_data])

    my_scada_data = ScadaData(sensor_config=my_sensor_config,
			         frozen_sensor_config=True,
                                 sensor_readings_time=my_measurement_time_points,
                                 pressure_data_raw=one_d_pressure_data,
                                 # pressure_data_raw=two_d_pressure_data,
                                 )
```

throws errors my way so I don't think I'm getting the pressure data format correct!
