### 2024-11-08 WaterFutures/EPyT-Flow#9 @meghnathomas
Just to add to my previous comment - the suggestion to add the additional example is totally optional and I think this paper definitely merits publication in JOSS without it, especially as you satisfy your goals to provide WDS data generation capabilities and a platform to host models and benchmark networks and scenarios. I only suggest its addition in the eventuality that researchers want to move beyond testing on benchmark networks/models/scenarios and begin to test their control strategies and analysis approaches on real systems and data.  I have no further comments beyond this and am happy to recommend this paper for acceptance. I will open a separate issue outside of the JOSS review requesting the example at a later date whenever convenient for the authors :-)
