### 2024-11-08 e87064e
ScadaData: Update documentation
