### 2024-11-08 9d13280
ScenarioSimulator: Separate function for hydraulic only simulation
