+++ epyt_flow/simulation/scenario_simulator.py
@@ -1644,64 +1644,73 @@ class ScenarioSimulator():
 
         if reporting_time_start == 0:
             if return_as_dict is True:
-                yield {"bulk_species_node_concentration_raw": bulk_species_node_concentrations,
-                       "bulk_species_link_concentration_raw": bulk_species_link_concentrations,
-                       "surface_species_concentration_raw": surface_species_concentrations,
-                       "sensor_readings_time": np.array([0])}
+                data = {"bulk_species_node_concentration_raw": bulk_species_node_concentrations,
+                        "bulk_species_link_concentration_raw": bulk_species_link_concentrations,
+                        "surface_species_concentration_raw": surface_species_concentrations,
+                        "sensor_readings_time": np.array([0])}
             else:
-                yield ScadaData(sensor_config=self.__sensor_config,
-                                bulk_species_node_concentration_raw=bulk_species_node_concentrations,
-                                bulk_species_link_concentration_raw=bulk_species_link_concentrations,
-                                surface_species_concentration_raw=surface_species_concentrations,
-                                sensor_readings_time=np.array([0]),
-                                sensor_reading_events=self.__sensor_reading_events,
-                                sensor_noise=self.__sensor_noise,
-                                frozen_sensor_config=frozen_sensor_config)
+                data = ScadaData(sensor_config=self.__sensor_config,
+                                 bulk_species_node_concentration_raw=bulk_species_node_concentrations,
+                                 bulk_species_link_concentration_raw=bulk_species_link_concentrations,
+                                 surface_species_concentration_raw=surface_species_concentrations,
+                                 sensor_readings_time=np.array([0]),
+                                 sensor_reading_events=self.__sensor_reading_events,
+                                 sensor_noise=self.__sensor_noise,
+                                 frozen_sensor_config=frozen_sensor_config)
+
+            if support_abort is True:  # Can the simulation be aborted? If so, handle it.
+                abort = yield
+                if abort is True:
+                    return None
+
+            yield data
 
         # Run step-by-step simulation
         tleft = 1
         total_time = 0
         while tleft > 0:
-            if support_abort is True:  # Can the simulation be aborted? If so, handle it.
-                abort = yield
-                if abort is not False:
-                    break
-
             # Compute current time step
             total_time, tleft = self.epanet_api.stepMSXQualityAnalysisTimeLeft()
 
-            if verbose is True:
-                try:
-                    next(progress_bar)
-                except StopIteration:
-                    pass
-
             # Fetch data at regular time intervals
             if total_time % hyd_time_step == 0:
+                if verbose is True:
+                    try:
+                        next(progress_bar)
+                    except StopIteration:
+                        pass
+
                 bulk_species_node_concentrations, bulk_species_link_concentrations, \
                     surface_species_concentrations = __get_concentrations()
 
                 # Report results in a regular time interval only!
                 if total_time % reporting_time_step == 0 and total_time >= reporting_time_start:
                     if return_as_dict is True:
-                        yield {"bulk_species_node_concentration_raw":
-                                   bulk_species_node_concentrations,
-                               "bulk_species_link_concentration_raw":
-                                   bulk_species_link_concentrations,
-                               "surface_species_concentration_raw": surface_species_concentrations,
-                               "sensor_readings_time": np.array([total_time])}
+                        data = {"bulk_species_node_concentration_raw":
+                                    bulk_species_node_concentrations,
+                                "bulk_species_link_concentration_raw":
+                                    bulk_species_link_concentrations,
[…]
