### 2024-11-08 1c6def5
Use stable doc in releases
