### 2024-11-10 #7104 @cheginit
Please address the following in the paper:

- [ ] L35-38: This sentence is long and difficult to read. Please break it down into sentences. Considering its previous sentence, it's kinda redundant. You can shorten it by removing the parts that repeat the previous sentence.
- [ ] L53: "a" is not needed in "such a data".
- [ ] L56-57: This sentence: "simulation of hydraulic dynamics, and in addition, it allows the simulation of various events such as" is awkward. You can rewrite it like so: "performing hydraulic simulations under various scenarios such as".
- [ ] L62-63: These two citations need to be in a single parenthesis using `[@author1:2001; @author2:2001]`.
- [ ] L67: Remove "the".
- [ ] L67-70: This sentence is long, please break it down into two sentences.
- [ ] L88: "and" is missing before "water quality"
- [ ] L101: Replace "&" with a comma
- [ ] L118: Capitalize Python

Once you're done with addressing these issues, please release a new version and with the new version create an archive, for example, on Zenodo, and provide me with both version number and DOI.
