### 2024-11-10 892fe4b
Bugfix frozen sensor config in ScadaData
