### 2024-11-10 #7104 @andreArtelt
Zenodo entry: https://zenodo.org/records/14063464
