### 2024-11-10 4747978
JOSS submission: Revise paper according to editor suggestions
