+++ paper/paper.md
@@ -56,7 +56,7 @@ To accelerate research in WDN management, `EPyT-Flow` provides easy access to po
 Water Distribution Networks (WDNs) are designed to ensure a reliable supply of drinking water. These systems are operated and monitored by humans, supported by software tools, including basic control algorithms and event detectors that rely on a limited number of sensors within the WDN. These sensors measure hydraulic (e.g., pressure, flow) and water quality (e.g., chemical concentrations) states. However, given the rapid population growth of urban areas, WDNs are becoming more complex to manage due to the resulting time-varying system uncertainty. Consequently, key tasks such as event detection (e.g., leakage) and isolation, pump scheduling, and control are becoming more challenging. Moreover, modeling and predicting water quality in the distribution network is becoming more difficult due to changing environmental conditions. This is why water utilities are now driven to install even more sensors to gather data on their changing systems. Traditionally, model-based methods were used for planning and managing WDNs; however, due to rapid changes, these methods may no longer be sufficient. New AI and data-driven methods can now take advantage of big data and are promising tools for tackling challenges in water management.
 
 Currently, non-water experts such as AI researchers face several challenges when devising practical solutions for water system applications, such as the unavailability of tools for easy scenario/data generation and easy access to benchmarks, which hinder the progress of applying AI to this domain. 
-Easy-to-use toolboxes and access to benchmark data sets are extremely important for boosting and accelerating research, as well as for supporting reproducible research, as it was, for instance, the case in deep learning and machine learning where toolboxes such as [TensorFlow](https://www.tensorflow.org/) and [scikit-learn](https://scikit-learn.org/stable/) had a significant impact on boosting research.
+Easy-to-use toolboxes and access to benchmark data sets are extremely important for boosting and accelerating research, as well as for supporting reproducible research. This was, for instance, the case in deep learning and machine learning with toolboxes such as [TensorFlow](https://www.tensorflow.org/) and [scikit-learn](https://scikit-learn.org/stable/).
 
 
 ## State of the field
@@ -73,29 +73,29 @@ in WDN resilience and response to various operational challenges.
 These tools, however, do not support the creation of realistic (benchmark) scenarios with
 essential aspects such as realistic fault models (of leakages and sensor faults), various sensor configurations,
 custom control modules, and other events such as changes in water quality caused by external factors.
-However, such a data creation mechanisms are important for supporting and enabling research on the application of
+However, such data creation mechanisms are important for supporting and enabling research on the application of
 Machine Learning (ML) and Artifical Intelligence (AI) in WDNs.
 A first step towards such software for scenario creation is the `Water Network Tool for Resilience (WNTR)` [@klise2017software],
-which facilitates the simulation of hydraulic dynamics, and in addition, it allows the simulation of various events
-such as pipe breaks, disasters such as earthquakes, power outages, and fires. However, it currently does not fully integrate
+which facilitates the performing hydraulic simulations under various scenarios such as
+pipe breaks, disasters such as earthquakes, power outages, and fires. However, it currently does not fully integrate
 (advanced) quality dynamics with scenarios such as pipe leaks and also misses other crucial modules such as
 sensor configurations, and considerations of industrial controls
 (i.e. controls that go beyond simple IF-THEN-ELSE controls as supported by EPANET).
 
-The transition to Python-based open-source software [@kyriakou2023epyt],[@klise2017software], underscores a broader trend
[…]
