### 2024-11-10 77164db
JOSS submission: Capitalize "Python" in references
