+++ paper/paper.bib
@@ -57,7 +57,7 @@
 }
 
 @article{kyriakou2023epyt,
-  title={EPyT: An EPANET-Python Toolkit for Smart Water Network Simulations},
+  title={{EPyT: An EPANET-Python Toolkit for Smart Water Network Simulations}},
   author={Kyriakou, Marios S and Demetriades, Marios and Vrachimis, Stelios G and Eliades, Demetrios G and Polycarpou, Marios M},
   journal={Journal of Open Source Software},
   volume={8},
