### 2024-11-10 #7104 @andreArtelt
Release: 0.7.2
DOI of Zenodo entry: 10.5281/zenodo.14063572
