### 2024-11-11 8d141c9
Merged with scenario visualizer
