+++ epyt_flow/simulation/scenario_visualizer.py
@@ -1,28 +1,1316 @@
 """
 Module provides a class for visualizing scenarios.
 """
+from typing import Optional, Union, List, Tuple, Iterable
+
 import matplotlib.pyplot as plt
+from matplotlib.animation import FuncAnimation
+import matplotlib as mpl
+import networkx.drawing.nx_pylab as nxp
+import numpy as np
+from svgpath2mpl import parse_path
 
 from .scenario_simulator import ScenarioSimulator
+from .scada.scada_data import ScadaData
+
+PUMP_PATH = ('M 202.5 93 A 41.5 42 0 0 0 161 135 A 41.5 42 0 0 0 202.5 177 A '
+             '41.5 42 0 0 0 244 135 A 41.5 42 0 0 0 241.94922 122 L 278 122 '
+             'L 278 93 L 203 93 L 203 93.011719 A 41.5 42 0 0 0 202.5 93 z')
+RESERVOIR_PATH = ('M 325 41 A 43 24.5 0 0 0 282.05664 65 L 282 65 L 282 65.5 '
+                  'L 282 163 L 282 168 L 282 216 L 305 216 L 305 168 L 345 '
+                  '168 L 345 216 L 368 216 L 368 168 L 368 163 L 368 65.5 L '
+                  '368 65 L 367.98047 65 A 43 24.5 0 0 0 325 41 z')
+TANK_PATH = ('M 325 41 A 43 24.5 0 0 0 282.05664 65 L 282 65 L 282 65.5 L 282 '
+             '185 L 368 185 L 368 65.5 L 368 65 L 367.98047 65 A 43 24.5 0 0'
+             ' 0 325 41 z')
+VALVE_PATH = ('M 9.9999064 9.9999064 L 9.9999064 110 L 69.999862 59.999955 L '
+              '9.9999064 9.9999064 z M 69.999862 59.999955 L 129.99982 110 L '
+              '129.99982 9.9999064 L 69.999862 59.999955 z')
 
 
-class ScenarioVisualizer():
+class Marker:
     """
-    Class for visualizing a given scenario.
+    The Marker class provides svg representations of hydraulic components
+    (pump, reservoir, tank and valve), which are loaded from their respective
+    svg paths and transformed into matplotlib.path.Path objects in order to be
+    used with the matplotlib library.
 
-    Parameters
+    Attributes
     ----------
-    scenario : :class:`~epyt_flow.simulation.scenario_simulator.ScenarioSimulator`
-        Scenario to be visualized.
+    pump : matplotlib.path.Path object
+        Marker for the pump, loaded from PUMP_PATH.
+    reservoir : matplotlib.path.Path object
+        Marker for the reservoir, loaded from RESERVOIR_PATH.
+    tank : matplotlib.path.Path object
+        Marker for the tank, loaded from TANK_PATH.
+    valve : matplotlib.path.Path object
+        Marker for the valve, loaded from VALVE_PATH.
+
+    Methods
+    -------
+    __marker_from_path(path, scale_p=1)
+        Loads and applies transformations to the marker shape from the given
+        path.
+    """
+    def __init__(self):
+        """
+        Initializes the Marker class and assigns matplotlib.path.Path markers
+        for pump, reservoir, tank, and valve components.
+        """
+        self.pump = self.__marker_from_path(PUMP_PATH, 2)
+        self.reservoir = self.__marker_from_path(RESERVOIR_PATH)
+        self.tank = self.__marker_from_path(TANK_PATH)
+        self.valve = self.__marker_from_path(VALVE_PATH)
+
+    @staticmethod
+    def __marker_from_path(path, scale_p=1):
+        """
+        Loads the marker from the specified path and adjusts it representation
+        by aligning, rotating and scaling it.
+
+        Parameters
+        ----------
+        path : str
[…]
