+++ epyt_flow/simulation/scenario_visualizer.py
@@ -32,18 +32,18 @@ class Marker:
     """
     The Marker class provides svg representations of hydraulic components
     (pump, reservoir, tank and valve), which are loaded from their respective
-    svg paths and transformed into matplotlib.path.Path objects in order to be
-    used with the matplotlib library.
+    svg paths and transformed into :class:`~matplotlib.path.Path` objects in
+    order to be used with the matplotlib library.
 
     Attributes
     ----------
-    pump : matplotlib.path.Path object
+    pump : :class:`~matplotlib.path.Path` object
         Marker for the pump, loaded from PUMP_PATH.
-    reservoir : matplotlib.path.Path object
+    reservoir : :class:`~matplotlib.path.Path` object
         Marker for the reservoir, loaded from RESERVOIR_PATH.
-    tank : matplotlib.path.Path object
+    tank : :class:`~matplotlib.path.Path` object
         Marker for the tank, loaded from TANK_PATH.
-    valve : matplotlib.path.Path object
+    valve : :class:`~matplotlib.path.Path` object
         Marker for the valve, loaded from VALVE_PATH.
 
     Methods
@@ -54,8 +54,8 @@ class Marker:
     """
     def __init__(self):
         """
-        Initializes the Marker class and assigns matplotlib.path.Path markers
-        for pump, reservoir, tank, and valve components.
+        Initializes the Marker class and assigns :class:`~matplotlib.path.Path`
+        markers for pump, reservoir, tank, and valve components.
         """
         self.pump = self.__marker_from_path(PUMP_PATH, 2)
         self.reservoir = self.__marker_from_path(RESERVOIR_PATH)
@@ -63,21 +63,21 @@ class Marker:
         self.valve = self.__marker_from_path(VALVE_PATH)
 
     @staticmethod
-    def __marker_from_path(path, scale_p=1):
+    def __marker_from_path(path: str, scale_p: int = 1) -> mpl.path.Path:
         """
         Loads the marker from the specified path and adjusts it representation
         by aligning, rotating and scaling it.
 
         Parameters
         ----------
-        path : str
+        path : `str`
             The svg path describing the marker shape.
-        scale_p : float, optional
+        scale_p : `float`, optional
             Scaling factor for the marker (default is 1).
 
         Returns
         -------
-        marker_tmp : matplotlib.path.Path object
+        marker_tmp : :class:`~matplotlib.path.Path` object
             The transformed marker object after loading and adjusting it.
         """
         marker_tmp = parse_path(path)
@@ -94,47 +94,48 @@ class ScenarioVisualizer:
     This class provides the necessary function to generate visualizations in
     the form of plots or animations from water network data.
 
-    Given a ScenarioSimulator object, this class provides the necessary
-    functions to plot the network topology and to color hydraulic elements
-    according to simulation data.
-    The resulting plot can then either be displayed or saved.
+    Given a :class:`~epyt_flow.simulation.ScenarioSimulator` object, this class
+    provides the necessary functions to plot the network topology and to color
+    hydraulic elements according to simulation data. The resulting plot can
+    then either be displayed or saved.
 
     Attributes
     ----------
-    __scenario : epyt_flow.simulation.ScenarioSimulator
+    __scenario : :class:`~epyt_flow.simulation.ScenarioSimulator`
         ScenarioSimulator object containing the network topology and
         configurations to obtain the simulation data which should be displayed.
-    fig : matplotlib.pyplot.Figure or None
+    fig : :class:`~matplotlib.pyplot.Figure` or None
         Figure object used for plotting, created and customized by calling the
         methods of this class, initialized as None.
-    ax : matplotlib.axes.Axes or None
[…]
