### 2024-11-11 #7104 @cheginit
@andreArtelt thanks for addressing the comments. However, there seems to be a misunderstanding, when I said capitalize Python, I meant only the first letter and for Line 118, not throughout capitalizing all letters throughout the paper. Please consistently use Python everywhere. Also, for the "EPyT: An EPANET-Python Toolkit for Smart Water Network Simulations" item in references, please capitalize only the first word to be consistent with the other items, i.e., "EPyT: An EPANET-Python toolkit for smart water network simulations"
