### 2024-11-11 c5bbea1
JOSS submission: Revise references according to editor's suggestions
