### 2024-11-11 66ee7ee
JOSS submission: Capitalize "Python" according to editor's suggestions
