### 2024-11-12 8e86aa0
Bug fix - would not display pit 0
